+++ tests/test_notify_import_safety.py
@@ -127,6 +127,8 @@ def test_notification_kinds_taxonomy_intact():
         "prep_failure",
         "interview_prep_ready",
         "interview_prep_failed",
+        "study_guide_failed",
+        "flashcard_failed",
         "recall_audit_alert",
         "drift_alert",
     }